fix: restructure test suite to avoid to many external dependencies
The intention behind this is to make tests faster and more robust. Especially to avoid false positives where the test works, but fails due to interaction with HF, the OS (CLI commands).

- Restructure test suite to make it easier to find specific tests and add tests in the future. e.g. added all evaluators test to a folder
- splits test_mteb into multiple segments, one for testing SentenceTransformers, one for testing evaluation workflow, one for testing integration with datasets
- avoid using encoders when testing evaluation workflow (one extra interaction with hf which can fail and it is slow)
- Added MockEncoders for testing torch, numpy and torch bf16 return types (see FP16 embeddings fail with BitextMining #1064)
- Added MockTask for all task types to avoid downloading too many datasets and to allow for more diverse without making the test suite too slow

The slowest test are now the mteb_rerank tests along with the integration tests. @orionw was hoping we might be able to to replace the dataset with a dummy dataset. However the slowest atm. is 20s

This PR also added a few minor fixes (surprisingly few actually) for handling torch bf16 which was not handled and a minor fix to AbstaskClusteringFast (@isaac-chung)

+++ mteb/evaluation/evaluators/model_encode.py
@@ -39,6 +39,6 @@ def model_encode(
 
     embeddings = model.encode(sentences, **kwargs)
     if isinstance(embeddings, torch.Tensor):
-        embeddings = embeddings.cpu().detach()
+        embeddings = embeddings.cpu().detach().float()
 
     return np.asarray(embeddings)
